finish PostForMainPageManagerImplTest.java

+++ src/main/java/com/ita/edu/softserve/manager/impl/PostForMainPageManagerImpl.java
@@ -9,7 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.ita.edu.softserve.dao.impl.PostDAOImpl;
 import com.ita.edu.softserve.entity.Post;
-import com.ita.edu.softserve.manager.ManagerFactory;
 import com.ita.edu.softserve.manager.PostForMainPageManager;
 import com.ita.edu.softserve.manager.UserNameService;
 
@@ -32,44 +31,46 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 
 	@Autowired
 	private PostDAOImpl postDao;
-	
+
 	@Autowired
 	private UserNameService userName;
 
 	@Transactional(readOnly = true)
 	@Override
 	public List<Post> findPostList() {
-		
+
 		try {
-			
+
 			return postDao.getAllEntities();
 		} catch (RuntimeException e) {
-			LOGGER.error(findPostMsg , e);
+			LOGGER.error(findPostMsg, e);
 			throw e;
 		}
-	
+
 	}
 
 	@Transactional
 	@Override
-	public boolean createNews(String newsTitle, String newsDescription, String imageSrc) {
-		Post post = null ;
+	public boolean createNews(String newsTitle, String newsDescription,
+			String imageSrc) {
+		Post post = null;
 		post = postDao.findByTitle(newsTitle);
-		
+
 		if (post == null) {
-		try {
-			
-			post = new Post(newsTitle, newsDescription, imageSrc);
-			postDao.save(post);
-			
-			LOGGER.info(entityName + post.getPostId() + addMsg + userName.getLoggedUsername());
-			
-			return true;
-		
-		} catch (RuntimeException e) {
-			LOGGER.error(createPostMsg,e);
-			return false;
-		}
+			try {
+
+				post = new Post(newsTitle, newsDescription, imageSrc);
+				postDao.save(post);
+
+				LOGGER.info(entityName + post.getPostId() + addMsg
+						+ userName.getLoggedUsername());
+
+				return true;
+
+			} catch (RuntimeException e) {
+				LOGGER.error(createPostMsg, e);
+				throw e;
+			}
 		}
 		return false;
 	}
@@ -77,28 +78,39 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 	@Transactional
 	@Override
 	public boolean removeNews(Integer id) {
-		try {
-			Post post = postDao.findById(id);
-			postDao.remove(post);
-			LOGGER.info(entityName + id + removeMsg + userName.getLoggedUsername());
-			return true;
-		} catch (RuntimeException e) {
-			LOGGER.error(removePostMsg ,e);
+		Post post = null;
+		post = postDao.findById(id);
+		if (post != null) {
+			try {
+				postDao.remove(post);
+				LOGGER.info(entityName + id + removeMsg
+						+ userName.getLoggedUsername());
+				return true;
+			} catch (RuntimeException e) {
+				LOGGER.error(removePostMsg, e);
+				throw e;
+			}
 		}
 		return false;
 	}
-	
+
 	@Transactional
 	@Override
 	public boolean removePost(String title) {
-		try {
-			Post post = postDao.findByTitle(title);
-			int id = post.getPostId();
-			postDao.remove(post);
-			LOGGER.info(entityName + id + removeMsg + userName.getLoggedUsername());
-			return true;
-		} catch (RuntimeException e) {
-			LOGGER.error(removePostMsg ,e);
+		Post post = null;
+		post = postDao.findByTitle(title);
+
+		if (post != null) {
+			try {
+				int id = post.getPostId();
+				postDao.remove(post);
+				LOGGER.info(entityName + id + removeMsg
+						+ userName.getLoggedUsername());
+				return true;
+			} catch (RuntimeException e) {
+				LOGGER.error(removePostMsg, e);
+				throw e;
+			}
 		}
 		return false;
 	}
@@ -106,7 +118,7 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 	@Transactional(readOnly = true)
 	@Override
 	public Post findNews(Integer postId) {
-		
+
 		try {
 			Post post = postDao.findById(postId);
 			return post;
@@ -120,21 +132,24 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 	@Override
 	public boolean updateNews(Integer newsId, String newsTitle,
 			String newsDescription, String imageSrc) {
+		Post post = null;
+		post = postDao.findById(newsId);
 
-		try {
-			Post post = postDao.findById(newsId);
-			if (post != null) {
-			post.setTitle(newsTitle);
-			post.setDescription(newsDescription);
-			post.setDate();
-			post.setImgSrc(imageSrc);
-			postDao.update(post);
-			LOGGER.info(entityName + post.getPostId() + changeMsg + userName.getLoggedUsername());
-			
-			return true;
+		if (post != null) {
+			try {
+				post.setTitle(newsTitle);
+				post.setDescription(newsDescription);
+				post.setDate();
+				post.setImgSrc(imageSrc);
+				postDao.update(post);
+				LOGGER.info(entityName + post.getPostId() + changeMsg
+						+ userName.getLoggedUsername());
+
+				return true;
+			} catch (RuntimeException e) {
+				LOGGER.error(updatePostMsg, e);
+				throw e;
 			}
-		} catch (RuntimeException e) {
-			LOGGER.error(updatePostMsg ,e);
 		}
 		return false;
 	}
@@ -144,7 +159,7 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 		try {
 			return postDao.getPostListCount();
 		} catch (RuntimeException e) {
-			LOGGER.error(countPostMsg,e);
+			LOGGER.error(countPostMsg, e);
 			throw e;
 		}
 	}
@@ -154,13 +169,9 @@ public class PostForMainPageManagerImpl implements PostForMainPageManager {
 		try {
 			return postDao.getPostForOnePage(from, count);
 		} catch (RuntimeException e) {
-			LOGGER.error(resultPerPagePostMsg,e);
+			LOGGER.error(resultPerPagePostMsg, e);
 			throw e;
 		}
-	}
-
-	public static PostForMainPageManager getInstance() {
-		return ManagerFactory.getManager(PostForMainPageManager.class);
 	}
 
 }
